# Optimize --find-links=<path-to-dir>