Fixes https://github.com/pypa/pip/issues/12320 by implementing the optimization strategy:

* All instances of `_FlatDirectorySource` only scan the relevant path once (instead of currently which is 2*(number of projects pip considers installing)) by caching the result of scanning

Further optimizations were added:

* Remove `Path` construction for only interating directory, as the result was just stingified anyway
* `_FlatDirectorySource` only considers file names in the path that have a valid sdist of wheel name
* `_FlatDirectorySource` only considers file names that are relevant to the current project Pip is considering, which produces a huge speed up where there are a lot of projects to be installed

To make sure this optimization did not produce unintended slowdowns I tried 4 different requirements on my machine using the steps outlined in https://github.com/pypa/pip/issues/12320, taking the median of each run I saw all 4 saw performance improvements (although for a small number of wheels it was small enough to be negligible):

| Requirement | Pip Main Median Time | Pip Optimized Median Time | Relative Performance Improvement |
| --- | --- | --- | --- |
| `requests` | 0.3 seconds | 0.296 seconds | 1.3% |
| `pandas[performance,excel,html,parquet]` | 0.662s seconds | 0.645s seconds | 2.6% |
| `apache-airflow[all]` | 45.633 seconds | 30.724 seconds | 32.7% |
| `homeassistant` from https://github.com/pypa/pip/issues/12320  | 159.206 seconds | 68.969 seconds | 56.7% |