@notatallshaw 2023-10-08
Optimizations strategies in this PR might be able to be applied to other ways Pip collects links. I will take a look at them when I have a way to consistently benchmarking those.