@notatallshaw 2023-10-08
Profiling the `homeassistant` usage show that in main 47% of the time was taken while under the function `pip._internal.index.package_finder.find_best_candidate`, under this optimized branch that is reduced to 0.66% of the time.

The biggest hotspot under this branch for that test case is now `pip._vendor.resolvelib.resolvers._is_current_pin_satisfying` which is being called millions of times by `pip._vendor.resolvelib.resolvers.resolve`, if this can somehow be optimized it would have performance gains for almost all Pip installs.