@notatallshaw 2023-10-19
Updated to significantly simplify the new class `_FlatDirectoryToUrls`, rather than it implementing some singleton-like behavior the class `_FlatDirectorySource` just keeps a class level variable keeping track of what `_FlatDirectoryToUrls` classes have already been instantiated and can be reused.

Also cleaned the commit history so it's just 3 commits.