@notatallshaw 2023-10-19
> I do intend to try to review this at some point, but life has been hectic, so I can't promise when I'll be able to get to it.

Absolutely no rush, I don't consider this PR particularly critical in of itself.

But I planned to benchmark Pip's other ways of collecting packages and seeing if I can reuse the same approaches to improve their performances as well. And it also it helps surface other `O(n^2)` issues while profiling Pip collecting lots of packages.