@notatallshaw 2024-01-30
Thanks for merging, I will keep an eye out on the issue tracker after the next Pip release, this will probably have an unintended effect on someone's unusual workflow.